Update to use new features from dev
