+++ src/pages/Signup.jsx
@@ -1,12 +1,18 @@
 // React Imports
 import React, { useState, useEffect } from 'react';
-import { TextField, Button, CardHeader, textFieldClasses } from '@mui/material';
+import { TextField, Button, CardHeader } from '@mui/material';
 
 import {
-  login,
-  getDefaultSession,
-  handleIncomingRedirect
-} from '@inrupt/solid-client-authn-browser';
+  buildThing,
+  createThing,
+  mockSolidDatasetFrom,
+  saveSolidDatasetAt,
+  setThing
+} from '@inrupt/solid-client';
+
+import { RDF_PREDICATES } from '@constants';
+
+import { login, handleIncomingRedirect } from '@inrupt/solid-client-authn-browser';
 
 import { useSearchParams } from 'react-router-dom';
 import Box from '@mui/material/Box';
@@ -14,9 +20,6 @@ import Card from '@mui/material/Card';
 import Container from '@mui/material/Container';
 import Paper from '@mui/material/Paper';
 import Typography from '@mui/material/Typography';
-
-import { registerPod, subscribeToUser } from '../utils';
-import { OidcLoginComponent } from '../components/NavBar';
 
 /**
  * Signup - First screen in the user registration flow.
@@ -32,6 +35,50 @@ const formRowStyle = {
 
 const textFieldStyle = {
   margin: '8px'
+};
+
+export const registerPod = async (
+  { email, password, confirmPassword },
+  oidcProvider = import.meta.env.VITE_SOLID_POD_SERVER
+) => {
+  const [podName] = email.split('@');
+
+  const oidcRegistrationPath = `${oidcProvider}idp/register/`;
+
+  const body = {
+    email,
+    password,
+    confirmPassword,
+    podName,
+    createWebId: true,
+    createPod: true,
+    rootPod: false,
+    register: true
+  };
+
+  const response = await fetch(oidcRegistrationPath, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json'
+    },
+    body: JSON.stringify(body)
+  });
+  return response.json();
+};
+
+export const subscribeToUser = async (userPodUrl, myProfile) => {
+  const { myWebId, myPodUrl, myEmail } = myProfile;
+  const [username] = myEmail.split('@');
+  const thing = buildThing(createThing({ name: username }))
+    .addUrl(RDF_PREDICATES.identifier, myWebId)
+    .addUrl(RDF_PREDICATES.URL, myPodUrl)
+    .build();
+  const datasetUrl = `${userPodUrl}Users/userlist.ttl`;
+  // Inrupt's libraries don't seem to support append-only access to datasets normally.
+  // mockSolidDatasetFrom is a workaround
+  let dataset = mockSolidDatasetFrom(datasetUrl);
+  dataset = setThing(dataset, thing);
+  await saveSolidDatasetAt(datasetUrl, dataset);
 };
 
 const Signup = () => {
