+++ src/utils/index.js
@@ -6,13 +6,13 @@
  * @module utils
  * @namespace utils
  */
-import runNotification from './frontend/notification-helper';
 import getDriversLicenseData from './barcode/barcode-scan';
 import formattedDate from './barcode/barcode-date-parser';
 
-export { runNotification, getDriversLicenseData, formattedDate };
+export { getDriversLicenseData, formattedDate };
 
 export * from './cryptography/credentials-helper';
 export * from './network/session-core';
 export * from './network/session-helper';
 export * from './pod-management/pod-helper';
+export * from './frontend/notification-helper';
